+++ src/app/auth/models/user.model.ts
@@ -1,0 +1,7 @@
+export interface Authenticate {
+    userName: string;
+    password: string;
+    rememberMe: boolean;
+}
+
+