+++ src/app/auth/services/authentication.service.ts
@@ -1,0 +1,5 @@
+import { HttpModule } from '@angular/http';
+
+export class AuthenticationService {
+
+}