+++ src/app/core/containers/not-found-page.component.ts
@@ -1,0 +1,10 @@
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+    moduleId: module.id,
+    selector: 'not-found-page',
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    template: 'not-found-page.component.html'
+})
+
+export class NotFoundPageComponent { }