+++ src/app/main.ts
@@ -1,6 +1,6 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { enableProdMode } from '@angular/core';
-import { AppSettings } from './constants/AppSettings';
+import { AppSettings } from './shared/constants/AppSettings';
 
 if (AppSettings.ENVIRONMENT === 'prod') {
     enableProdMode();